@@ -1,12 +1,16 @@
 module github.com/drone-plugins/drone-curseforge
 
-go 1.13
+go 1.19
 
 require (
-	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
-	github.com/drone-plugins/drone-plugin-lib v0.3.0
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/sirupsen/logrus v1.4.2
-	github.com/urfave/cli/v2 v2.1.1
-	golang.org/x/sys v0.0.0-20200121082415-34d275377bf9 // indirect
+	github.com/drone-plugins/drone-plugin-lib v0.4.0
+	github.com/sirupsen/logrus v1.9.0
+	github.com/urfave/cli/v2 v2.23.7
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	golang.org/x/sys v0.3.0 // indirect
 )
